+++ Services/Interfaces/IAdmissionBatches.cs
@@ -6,10 +6,11 @@ namespace EduReg.Services.Interfaces
     public interface IAdmissionBatches
     {
         Task<GeneralResponse> CreateAdmissionBatchAsync(AdmissionBatchesDto model);
-        Task<GeneralResponse> UpdateAdmissionBatchAsync(int Id, AdmissionBatchesDto model);
-        Task<GeneralResponse> DeleteAdmissionBatchAsync(int Id);
-        Task<GeneralResponse> GetAdmissionBatchByIdAsync(int Id);
+        Task<GeneralResponse> UpdateAdmissionBatchAsync(long Id, AdmissionBatchesDto model);
+        Task<GeneralResponse> DeleteAdmissionBatchAsync(long Id);
+        Task<GeneralResponse> GetAdmissionBatchByIdAsync(long Id);
         Task<GeneralResponse> GetAllAdmissionBatchAsync(PagingParameters paging);
+
 
     }
 }
