+++ Services/Interfaces/IFeeItems.cs
@@ -7,9 +7,10 @@ namespace EduReg.Services.Interfaces
     {
         Task<GeneralResponse> CreateFeeItemAsync(FeeItemDto model);
         Task<GeneralResponse> AddFeeItemToSemesterScheduleAsync(List<FeeItemDto> model);
-        Task<GeneralResponse> UpdateFeeItemAsync(int Id, FeeItemDto model); 
-        Task<GeneralResponse> DeleteFeeItemAsync(int Id);
-        Task<GeneralResponse> GetFeeItemByIdAsync(int Id);
+        Task<GeneralResponse> UpdateFeeItemAsync(long Id, FeeItemDto model); 
+        Task<GeneralResponse> DeleteFeeItemAsync(long Id);
+        Task<GeneralResponse> GetFeeItemByIdAsync(long Id);
         Task<GeneralResponse> GetAllFeeItemsAsync(PagingParameters paging);
+
     }
 }
