+++ Services/Interfaces/IFeeRules.cs
@@ -7,9 +7,9 @@ namespace EduReg.Services.Interfaces
     {
         Task<GeneralResponse> CreateFeeRuleAsync(FeeRuleDto model, string institutionShortName);
         Task<GeneralResponse> ApplyFeeRuleToSemesterScheduleAsync(List<FeeRuleDto> models, string institutionShortName);
-        Task<GeneralResponse> UpdateFeeRuleAsync(int id, FeeRuleDto model, string institutionShortName);
-        Task<GeneralResponse> DeleteFeeRuleAsync(int id, string institutionShortName);
-        Task<GeneralResponse> GetFeeRuleByIdAsync(int id, string institutionShortName);
+        Task<GeneralResponse> UpdateFeeRuleAsync(long id, FeeRuleDto model, string institutionShortName);
+        Task<GeneralResponse> DeleteFeeRuleAsync(long id, string institutionShortName);
+        Task<GeneralResponse> GetFeeRuleByIdAsync(long id, string institutionShortName);
         Task<GeneralResponse> GetAllFeeRuleAsync(string institutionShortName, PagingParameters paging);
     }
 }
