+++ Services/Interfaces/ISemesters.cs
@@ -6,9 +6,9 @@ namespace EduReg.Services.Interfaces
     public interface ISemesters
     {
         Task<GeneralResponse> CreateSemesterAsync(SemestersDto model);
-        Task<GeneralResponse> UpdateSemesterAsync(int Id, SemestersDto model);
-        Task<GeneralResponse> DeleteSemesterAsync(int Id);
-        Task<GeneralResponse> GetSemesterByIdAsync(int Id);
+        Task<GeneralResponse> UpdateSemesterAsync(long Id, SemestersDto model);
+        Task<GeneralResponse> DeleteSemesterAsync(long Id);
+        Task<GeneralResponse> GetSemesterByIdAsync(long Id);
         Task<GeneralResponse> GetAllSemestersAsync(PagingParameters paging);
     }
 }
